# internal/server/openapi.yaml
openapi: 3.0.3
info:
  title: lux-api
  version: 0.1.0
  description: |
    HTTP wrapper around the [lux](https://github.com/iawia002/lux) video
    download library. Submit a URL to `POST /jobs`, then poll
    `GET /jobs/{id}` until `state == done` and fetch the file from
    `GET /jobs/{id}/file` — which redirects (`302`) to a short-lived
    presigned URL pointing at the MinIO/S3 object. Job state is stored in
    Postgres and survives restarts.

    Cancellation (`DELETE /jobs/{id}`) is best-effort: lux's downloader does
    not accept a `context.Context`, so an in-flight HTTP transfer will end on
    its own before the job is marked `canceled`.
servers:
  - url: ${LUX_SERVER_URL}
paths:
  /healthz:
    get:
      summary: Liveness check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }

  /jobs:
    get:
      summary: List all jobs
      responses:
        '200':
          description: Array of jobs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Job' }
    post:
      summary: Start a new download job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobRequest' }
            example:
              url: https://www.youtube.com/watch?v=dQw4w9WgXcQ
              stream: ""
              playlist: false
      responses:
        '202':
          description: |
            Job accepted. Streams have already been resolved synchronously; the
            response body is the initial Job snapshot (state=`queued` or
            `downloading`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '400': { $ref: '#/components/responses/Error' }
        '502':
          description: Extraction failed (URL unsupported, network error, site
            changed, etc.). No job is created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }

  /jobs/{id}:
    parameters:
      - $ref: '#/components/parameters/JobID'
    get:
      summary: Get job status (poll this for progress)
      responses:
        '200':
          description: Current job snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404':
          $ref: '#/components/responses/Error'
    delete:
      summary: Cancel a job and remove its partial files
      responses:
        '204':
          description: Canceled
        '404':
          $ref: '#/components/responses/Error'

  /jobs/{id}/file:
    parameters:
      - $ref: '#/components/parameters/JobID'
    get:
      summary: Redirect to a presigned MinIO URL for the result (only when state=done)
      responses:
        '302':
          description: Redirect to presigned download URL
          headers:
            Location:
              schema: { type: string, format: uri }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }

  /extract:
    post:
      summary: Resolve stream metadata without downloading
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobRequest' }
      responses:
        '200':
          description: lux extractor output
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
        '400': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }

components:
  parameters:
    JobID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }

  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }

  schemas:
    JobRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Source page URL recognised by a lux extractor
        stream:
          type: string
          description: Stream key (omit to pick the largest stream)
        playlist:
          type: boolean
          default: false
        cookie:
          type: string
          description: Cookie string forwarded to the extractor (for gated content)

    Job:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        state:
          type: string
          enum: [queued, extracting, downloading, uploading, done, failed, canceled]
        percent: { type: number, format: float, minimum: 0, maximum: 100 }
        bytes_downloaded: { type: integer, format: int64 }
        bytes_total: { type: integer, format: int64 }
        title: { type: string }
        site: { type: string }
        stream: { type: string }
        output_filename: { type: string }
        output_object_key: { type: string, description: "MinIO/S3 object key" }
        error: { type: string }
        playlist: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }